refactor: import providers by concrete submodule, not the package __init__
The providers import chain (providers/*.py + wiring.py) imported siblings from
the still-executing modern_di.providers package __init__, which worked only
because of the statement order there. Import concrete submodules instead, and
add an AST guard test pinning the invariant.
