+++ planning/changes/2026-07-13.01-providers-concrete-imports.md
@@ -1,0 +1,56 @@
+---
+summary: The providers import chain (providers/*.py + wiring.py) imports siblings by concrete submodule instead of from the still-executing modern_di.providers package __init__, so the package's statement order is no longer load-bearing; an AST guard test pins the invariant.
+---
+
+# Change: Import providers by concrete submodule, not the package __init__
+
+**Lane:** lightweight — four one-line import swaps plus one guard test. Each
+edit is mechanical; kept in this lane despite the file count because no logic,
+signature, or capability contract moves.
+
+## Goal
+
+Remove a latent import-order fragility: four modules that load *while*
+`modern_di/providers/__init__.py` is still executing reach back into the
+half-initialized package (`from modern_di.providers import ...`). It works only
+because `__init__` happens to bind `AbstractProvider`/`ContextProvider` before
+`factory`, and because the top-level `__init__` eagerly imports `container`
+first. Neither ordering is marked as load-bearing.
+
+## Approach
+
+Import siblings by their concrete module instead of the package:
+
+- `providers/container_provider.py`, `providers/context_provider.py` — `from
+  modern_di.providers import AbstractProvider` → `from
+  modern_di.providers.abstract import AbstractProvider`.
+- `providers/factory.py`, `wiring.py` — `from modern_di.providers import
+  ContextProvider` → `from modern_di.providers.context_provider import
+  ContextProvider`.
+
+External consumers outside the chain (e.g. `registries/cache_registry.py`) keep
+using the package API — they load after the package is fully initialized, so
+they are unaffected and unrestricted.
+
+A new AST guard in `tests/test_packaging.py` walks `providers/*.py` + `wiring.py`
+and fails on any `from modern_di.providers import ...` (level-0, package
+exactly), pinning the invariant so the fragility cannot silently return.
+
+No `architecture/` capability contract changes (this is import hygiene, not a
+behavior change). Sourced from candidate 2 of the 2026-07-12 architecture review.
+
+## Files
+
+- `modern_di/providers/container_provider.py` — concrete `AbstractProvider` import
+- `modern_di/providers/context_provider.py` — concrete `AbstractProvider` import
+- `modern_di/providers/factory.py` — concrete `ContextProvider` import
+- `modern_di/wiring.py` — concrete `ContextProvider` import
+- `tests/test_packaging.py` — AST guard test added
+
+## Verification
+
+- [x] Failing test first — guard reports `['container_provider.py:3', 'context_provider.py:6', 'factory.py:8', 'wiring.py:13']`.
+- [x] Apply the four import swaps.
+- [x] Guard test passes; `import modern_di.providers.factory` (submodule-first) succeeds.
+- [x] `just test-ci` — 359 passed, 100% coverage.
+- [x] `just lint-ci` — clean.
